TCPServer script feature, multithread encoding and sending, socket optimization, and some other bugs fixed: - New TCPServer feature in script (see "Server Application, Sockets" wiki page) - Optimization of reception sockets everywhere in the code (fast x2 on some location). - Multithread for encoding and sending process (CRC, AES, Pack, UDP::send) - A FlowNull crash bug fixed - ServerConnection inherits from Session and no more ServerSession - Some edges "died" logic solved

+++ CumulusLib/include/FlowNull.h
@@ -27,7 +27,7 @@ public:
 	FlowNull(Peer& peer,Invoker& invoker,BandWriter& band);
 	virtual ~FlowNull();
 	
-	void	fragmentHandler(Poco::UInt64 stage,Poco::UInt32 deltaNAck,PacketReader& fragment,Poco::UInt8 flags);
+	void	fragmentHandler(Poco::UInt64 stage,Poco::UInt64 deltaNAck,PacketReader& fragment,Poco::UInt8 flags);
 
 	FlowWriter&			writer();
