+++ CumulusLib/sources/FlowNull.cpp
@@ -33,7 +33,7 @@ FlowNull::FlowNull(Peer& peer,Invoker& invoker,BandWriter& band) : Flow(0,Util::
 FlowNull::~FlowNull() {
 }
 
-void FlowNull::fragmentHandler(UInt64 stage,UInt32 deltaNAck,PacketReader& fragment,UInt8 flags) {
+void FlowNull::fragmentHandler(UInt64 stage,UInt64 deltaNAck,PacketReader& fragment,UInt8 flags) {
 	fail("Message received for a Flow unknown");
 	(UInt64&)this->stage = stage;
 }
